Linkify a few plain URLs
The new Markdown parser – Kramdown – does not support autolinking plain
URLs in the text as far as I was able to learn. I used a rough macOS sed
to mark URLs as links. I ran the following command:

    for f in `ls _posts/*.md`; do sed -i .bak -E 's@ (https?://[^ ]+)( |[,.!?] )@ <\1>\2@g' "$f"; done; rm -f _posts/*.md.bak

Then viewed the diff and edited a few URLs that were incorrectly linked:

    git diff --word-diff _posts